feat: refactor into modular library and optimize quantization
- Restructured monolithic src/main.rs into src/lib.rs, src/config.rs, src/grid.rs, src/resample.rs, and src/error.rs.
- Replaced custom K-means with kmeans_colors crate using Lab color space for perceptual accuracy.
- Improved CLI with clap 4.x.
- Added comprehensive unit tests for core modules.
- Ensured WASM compatibility with wasm-pack.
- Fixed potential overflow and visibility issues.
