+++ Cargo.toml
@@ -5,13 +5,16 @@ edition = "2021"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
-path = "src/main.rs"
 
 [dependencies]
 image = { version = "0.24", default-features = false, features = ["png", "jpeg"] }
+clap = { version = "4.4", features = ["derive"] }
 rand = { version = "0.8", features = ["small_rng"] }
 rand_distr = "0.4"
 rand_chacha = "0.3"
+kmeans_colors = { version = "0.6", features = ["palette_color"] }
+palette = { version = "0.7", features = ["std"] }
+rayon = "1.8"
 
 # WASM-specific dependencies
 [target.'cfg(target_arch = "wasm32")'.dependencies]
