+++ src/error.rs
@@ -1,0 +1,39 @@
+use std::error::Error;
+use std::fmt;
+
+#[derive(Debug)]
+pub enum PixelSnapperError {
+    ImageError(image::ImageError),
+    InvalidInput(String),
+    ProcessingError(String),
+}
+
+impl fmt::Display for PixelSnapperError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            PixelSnapperError::ImageError(e) => write!(f, "Image error: {}", e),
+            PixelSnapperError::InvalidInput(msg) => write!(f, "Invalid input: {}", msg),
+            PixelSnapperError::ProcessingError(msg) => write!(f, "Processing error: {}", msg),
+        }
+    }
+}
+
+impl Error for PixelSnapperError {}
+
+impl From<image::ImageError> for PixelSnapperError {
+    fn from(error: image::ImageError) -> Self {
+        PixelSnapperError::ImageError(error)
+    }
+}
+
+#[cfg(target_arch = "wasm32")]
+use wasm_bindgen::prelude::*;
+
+#[cfg(target_arch = "wasm32")]
+impl From<PixelSnapperError> for JsValue {
+    fn from(err: PixelSnapperError) -> JsValue {
+        JsValue::from_str(&err.to_string())
+    }
+}
+
+pub type Result<T> = std::result::Result<T, PixelSnapperError>;
